+++ modules/core/src/main/scala/moe/brianhsu/live2d/enitiy/avatar/effect/impl/FaceTracking.scala
@@ -35,7 +35,14 @@ abstract class FaceTracking(protected val trackingTaps: TrackingTaps) extends Ef
     }
   }
 
-  private def calculateOperations(): List[ParameterValueUpdate] = {
+  private var lastFaceXAngle = 0.0f
+  private var lastFaceYAngle = 0.0f
+  private var lastFaceZAngle = 0.0f
+  private var lastBodyXAngle = 0.0f
+
+  private var isFirst = true
+
+  private def calculateOperations(): List[UpdateOperation] = {
     val faceXAngle = average(trackingNoes.take(trackingTaps.faceXAngle).map(_.faceXAngle))
     val faceYAngle = average(trackingNoes.take(trackingTaps.faceYAngle).map(_.faceYAngle))
     val faceZAngle = average(trackingNoes.take(trackingTaps.faceZAngle).map(_.faceZAngle))
@@ -46,10 +53,24 @@ abstract class FaceTracking(protected val trackingTaps: TrackingTaps) extends Ef
     val leftEyeSmile = average(trackingNoes.take(trackingTaps.leftEyeSmile).map(_.leftEyeSmile))
     val rightEyeSmile = average(trackingNoes.take(trackingTaps.rightEyeSmile).map(_.rightEyeSmile))
 
-    List(
-      ParameterValueUpdate("ParamAngleX", faceXAngle),
-      ParameterValueUpdate("ParamAngleY", faceYAngle),
-      ParameterValueUpdate("ParamAngleZ", faceZAngle),
+    if (isFirst) {
+      this.lastFaceXAngle = faceXAngle
+      this.lastFaceYAngle = faceYAngle
+      this.lastFaceZAngle = faceZAngle
+      this.lastBodyXAngle = faceXAngle
+      this.isFirst = false
+    }
+
+    this.lastFaceXAngle = lastFaceXAngle + calculateNewDiff(faceXAngle - lastFaceXAngle)
+    this.lastFaceYAngle = lastFaceYAngle + calculateNewDiff(faceYAngle - lastFaceYAngle)
+    this.lastFaceZAngle = lastFaceZAngle + calculateNewDiff(faceZAngle - lastFaceZAngle)
+    this.lastBodyXAngle = lastBodyXAngle + calculateNewDiff(faceXAngle - lastBodyXAngle, 0.85f)
+
+    val result = List(
+      ParameterValueUpdate("ParamAngleX", this.lastFaceXAngle),
+      ParameterValueUpdate("ParamAngleY", this.lastFaceYAngle),
+      ParameterValueUpdate("ParamAngleZ", this.lastFaceZAngle),
+      ParameterValueUpdate("ParamBodyAngleX", this.lastBodyXAngle, 0.75f),
       ParameterValueUpdate("ParamEyeLOpen", leftEyeOpenness),
       ParameterValueUpdate("ParamEyeROpen", rightEyeOpenness),
       ParameterValueUpdate("ParamMouthOpenY", mouthOpenness),
@@ -57,6 +78,22 @@ abstract class FaceTracking(protected val trackingTaps: TrackingTaps) extends Ef
       ParameterValueUpdate("ParamEyeLSmile", leftEyeSmile),
       ParameterValueUpdate("ParamEyeRSmile", rightEyeSmile),
     )
+
+
+    result
+  }
+
+  private def calculateNewDiff(originalDiff: Float, maxDiff: Float = 2.65f): Float = {
+
+    val afterLog = Math.log1p(originalDiff.abs).min(maxDiff).toFloat
+
+    if (originalDiff > 0) {
+      originalDiff.min(afterLog)
+    } else if (originalDiff < 0){
+      originalDiff.max(-afterLog)
+    } else {
+      originalDiff
+    }
   }
 
   private def average(values: List[Float]): Float = values.sum / values.size