+++ modules/core/src/main/scala/moe/brianhsu/live2d/exception/UnknownParameterTypeValueException.scala
@@ -1,0 +1,3 @@
+package moe.brianhsu.live2d.exception
+
+class UnknownParameterTypeValueException(value: Int) extends Exception(s"Unknown parameter type value: $value")