+++ modules/core/src/test/scala/moe/brianhsu/live2d/enitiy/avatar/motion/impl/AvatarMotionFeature.scala
@@ -3,7 +3,8 @@ package moe.brianhsu.live2d.enitiy.avatar.motion.impl
 import moe.brianhsu.live2d.adapter.gateway.avatar.motion.AvatarMotionDataReader
 import moe.brianhsu.live2d.adapter.gateway.avatar.settings.json.JsonSettingsReader
 import moe.brianhsu.live2d.enitiy.avatar.settings.Settings
-import moe.brianhsu.live2d.enitiy.model.{JavaVMParameter, Live2DModel}
+import moe.brianhsu.live2d.enitiy.model.Live2DModel
+import moe.brianhsu.live2d.enitiy.model.parameter.JavaVMParameter
 import moe.brianhsu.utils.expectation.ExpectedAvatarMotionOperation
 import moe.brianhsu.utils.expectation.ExpectedAvatarMotionOperation.Input
 import org.scalamock.scalatest.MockFactory